Controller Authorization of Update and Destroy

+++ app/Http/Controllers/UsersController.php
@@ -46,11 +46,15 @@ class UsersController extends Controller
     public function update(Request $request, $id)
     {
         $user = \App\User::find($id);
-        $user->user_id = \Auth::user()->id;
+        if ($user->user_id = \Auth::user()->id) {
         $user->name = $request->name;
         $user->email = $request->email;
         $user->password = $request->password;
         $user->save();
+
+        } else { 
+            return response ("Unauthorized", 403);
+        }
 
         return $user;
     }
@@ -64,7 +68,12 @@ class UsersController extends Controller
     public function destroy($id)
     {
         $user = \App\User::find($id);
+        if ($user->user_id = \Auth::user()->id) {
         $user->delete();
+
+        } else { 
+            return response ("Unauthorized", 403);
+        }
 
         return $user;
     }
